SCRD-1715 Support python3
Add support for python3 while retaining support for python2.7, following
the openstack guide at https://wiki.openstack.org/wiki/Python3.  That
guide explains the rationale behind nearly all of the changes in this
commit, including using six and oslo_serialization libraries, print
function, items function, utf8 encoding, and so on.

+++ ardana_service/main.py
@@ -144,7 +144,7 @@ def main():
                 "The default value, false, will cause calls to " \
                 "/api/v2/heartbeat to be rejected "
             LOG.error(msg)
-            print msg   # print on the console for good measure
+            print(msg)  # print on the console for good measure
             sys.exit(1)
 
         # Use our our middleware function to permit unsecured apis to be called
@@ -164,8 +164,9 @@ def main():
     app.config.from_mapping(flask_config)
 
     trigger_file = os.path.join(CONF.paths.log_dir, 'trigger.txt')
-    with open(trigger_file, 'w') as f:
-        f.write("Started at %s\n" % time.asctime())
+    if not os.path.exists(trigger_file):
+        with open(trigger_file, 'w') as f:
+            f.write("Started at %s\n" % time.asctime())
 
     socketio.init_app(app)
